feat: added first GH Action Workflow
This commit adds the first GH Action Workflow to the repository. This first workflow is responsible for scanning the repo for any secrets/vulnerable information using the gitleaks tool.

With this commit, the scan will run every day at 04:20 UTC, when a push is submitted, whenever a PR is opened or updated, and whenever a user manually requests one using the GitHub UI from the Actions tab.